requirements.txt
PyQt6
psutil
pymediainfo
imagehash
Pillow
geopy
